Reorganize client code, add BetaLive page
