+++ assets/js/beta/index.js
@@ -1,0 +1,8 @@
+const BetaHook = {
+  mounted() {
+    const betaNode = document.getElementById("beta");
+    if (betaNode) { console.log(`🧪 Beta mode enabled!`); }
+  }
+}
+
+export default BetaHook;
