--- lib/games_web/live/page_live.html.leex
@@ -1,1 +1,0 @@
-<div id="elm" phx-hook="elm"></div>
